ecom_cleaner/requirements.txt
streamlit==1.32.0
pandas==2.2.1
openpyxl==3.1.2
matplotlib==3.8.3
plotly==5.19.0
pyyaml==6.0.1
pyinstaller==6.4.0 